+++ flutter_inappwebview_web/lib/web/shims/platform_view_registry.dart
@@ -1,0 +1,2 @@
+export 'platform_view_registry_dart_ui.dart'
+    if (dart.library.ui_web) 'platform_view_registry_dart_ui_web.dart';