+++ flutter_inappwebview_web/lib/web/shims/platform_view_registry_dart_ui.dart
@@ -1,0 +1,10 @@
+import 'dart:html' as html;
+import 'dart:ui' as ui;
+
+class platformViewRegistry {
+  static bool registerViewFactory(
+      String viewTypeId, html.Element Function(int viewId) viewFactory) {
+    // ignore: undefined_prefixed_name
+    return ui.platformViewRegistry.registerViewFactory(viewTypeId, viewFactory);
+  }
+}